requirements.txt
python-dotenv==1.0.0
python-telegram-bot[rate-limiter]==20.4
peewee~=3.17.0
asyncpraw~=7.7.1
APScheduler==3.10.4
Pillow>=9.0.0
validators~=0.22.0
pymysql~=1.1.0
pytz~=2023.3
cryptography~=41.0.0
requests~=2.31.0
one-piece-wanted-poster>=0.0.12
geopy~=2.4.0
timezonefinder~=6.2.0
parsedatetime~=2.6
sentry-sdk~=1.39.1
black~=23.12.1